requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
pydantic

numpy
xgboost
lightgbm
catboost
nltk
textblob
tqdm


langchain 
faiss-cpu 
sentence-transformers 
langchain-community
langchain-openai
python-dotenv
langchain-huggingface
langchain-core
